pyproject.toml
[tool.poetry]
name = "parabank-pytest"
version = "0.1.0"
description = ""
authors = ["Pavlo Nahornyi <pavelnagornyj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.16.0"
pytest-playwright = "^0.4.3"
faker = "^22.5.1"
pytest-dotenv = "^0.5.2"
playwright = "^1.41.1"
pytest-html = "^4.1.1"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
